requirement.txt
PILLOW
screeninfo
pyautogui